Split Office Connector API tests to more files.
